♻️ Refactor iemapp tests into dedicated test file
